requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1
celery==5.3.4

# ML and Data Science
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.2
rdkit==2023.9.4

# Bioinformatics
biopython==1.82
pymol==2.5.0

# API and Web
httpx==0.26.0
beautifulsoup4==4.12.2
selenium==4.16.0

# Database
psycopg2-binary==2.9.9
motor==3.3.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Monitoring and Logging
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
